Move CRI internals to `cri.{cpp,h}`
